Updated to DXC6.6, newest Windows 10 SDK, and worked on low level D3D12 materials.

+++ Engine/Graphics/Direct3D12/D3D12CommonHeaders.h
@@ -9,6 +9,7 @@
 
 #include <dxgi1_6.h>
 #include <d3d12.h>
+//#include "H:\dev\HavanaEngine\Havana\packages\Microsoft.Direct3D.D3D12.1.608.2\build\native\include\d3d12.h"
 #include <wrl.h>
 
 #pragma comment(lib, "dxgi.lib")
@@ -48,14 +49,14 @@ if(FAILED(x)) {									\
 // Sets the name of a COM object and outputs a debug string into Visual Studio's output panel
 #define NAME_D3D12_OBJECT(obj, name) obj->SetName(name); OutputDebugString(L"::D3D12 Object Created: "); OutputDebugString(name); OutputDebugString(L"\n");
 // The indexed variant will include the index in the name of the object
-#define NAME_D3D12_OBJECT_INDEXED(obj, n, name)					\
-		{														\
-			wchar_t full_name[128];								\
-			if (swprintf_s(full_name, L"%s[%u]", name, n) > 0){	\
-				obj->SetName(full_name);						\
-				OutputDebugString(L"::D3D12 Object Created: ");	\
-				OutputDebugString(full_name);					\
-				OutputDebugString(L"\n");						\
+#define NAME_D3D12_OBJECT_INDEXED(obj, n, name)							\
+		{																\
+			wchar_t full_name[128];										\
+			if (swprintf_s(full_name, L"%s[%llu]", name, (u64)n) > 0){	\
+				obj->SetName(full_name);								\
+				OutputDebugString(L"::D3D12 Object Created: ");			\
+				OutputDebugString(full_name);							\
+				OutputDebugString(L"\n");								\
 		}}
 #else
 #define NAME_D3D12_OBJECT(x, name)
